+++ app/Controllers/HomeController.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Controllers;
+
+use App\Utils\View;
+
+class HomeController
+{
+    public function index()
+    {
+        echo View::render('home.index');
+    }
+}